agent/requirements.txt
pydantic-ai-slim[bedrock]
pydantic-ai-slim[logfire]
pydantic
boto3
logfire
phonenumbers